Move macro logic to utils to work with strings
