+++ crates/fixedpointmath/Cargo.toml
@@ -12,7 +12,7 @@ description = "Math library to simulate FixedPoint computation in Solidity smart
 [dependencies]
 
 # External dependencies
-ethers = "2.0.11"
+ethers = { version = "2.0.11", default-features = false }
 eyre = "0.6.8"
 rand = "0.8.5"
