+++ crates/fixedpointmath/src/macros.rs
@@ -11,44 +11,7 @@
 macro_rules! uint256 {
     ($number:expr) => {{
         let str = stringify!($number);
-
-        // Parse a string into a mantissa and an exponent. The U256 arithmetic
-        // will overflow if the mantissa or the exponent are too large.
-        let mut found_dot = false;
-        let mut found_e = false;
-        let mut mantissa = ethers::types::U256::zero();
-        let mut exponent = ethers::types::U256::zero();
-        let mut decimals = 0;
-
-        for digit in str.chars() {
-            if digit.is_ascii_digit() {
-                let d = digit.to_digit(10).unwrap();
-                if !found_e {
-                    mantissa = mantissa * 10 + d;
-                } else {
-                    exponent = exponent * 10 + d;
-                }
-                if found_dot && !found_e {
-                    decimals += 1;
-                }
-            } else if digit == 'e' && !found_e {
-                found_e = true;
-            } else if digit == '.' && !found_dot {
-                found_dot = true;
-            } else if digit != '_' {
-                panic!("Unexpected character: {digit}");
-            }
-        }
-
-        // Combine the mantissa and the exponent into a single U256. This will
-        // overflow if the exponent is too large. We also need to make sure that
-        // the final result is an integer.
-        let decimals = ethers::types::U256::from(decimals);
-        if exponent < decimals {
-            panic!("Exponent is too small: {exponent}");
-        }
-
-        mantissa * ethers::types::U256::from(10).pow(exponent - decimals)
+        $crate::u256_from_str(str).unwrap()
     }};
 }
 
@@ -56,46 +19,7 @@ macro_rules! uint256 {
 macro_rules! int256 {
     ($number:expr) => {{
         let str = stringify!($number);
-
-        // Parse a string into a mantissa and an exponent. The U256 arithmetic
-        // will overflow if the mantissa or the exponent are too large.
-        let mut sign = ethers::types::I256::one();
-        let mut found_dot = false;
-        let mut found_e = false;
-        let mut mantissa = ethers::types::I256::zero();
-        let mut exponent = 0;
-        let mut decimals = 0;
-
-        for digit in str.chars() {
-            if digit.is_ascii_digit() {
-                let d = digit.to_digit(10).unwrap();
-                if !found_e {
-                    mantissa = mantissa * 10 + d;
-                } else {
-                    exponent = exponent * 10 + d;
-                }
-                if found_dot && !found_e {
-                    decimals += 1;
-                }
-            } else if digit == '-' {
-                sign = -ethers::types::I256::one();
-            } else if digit == 'e' && !found_e {
-                found_e = true;
-            } else if digit == '.' && !found_dot {
-                found_dot = true;
-            } else if digit != '_' {
-                panic!("Unexpected character: {digit}");
-            }
-        }
-
-        // Combine the mantissa and the exponent I256. This will
-        // overflow if the exponent is too large. We also need to make sure that
-        // the final result is an integer.
-        if exponent < decimals {
-            panic!("Exponent is too small: {exponent}");
-        }
-
-        sign * mantissa * ethers::types::I256::from(10).pow(exponent - decimals)
+        $crate::i256_from_str(str).unwrap()
     }};
 }
